user-service/requirements.txt
asyncpg==0.20.1
databases[postgresql]==0.2.6
fastapi==0.48.0
SQLAlchemy==1.3.13
uvicorn==0.11.2
httpx==0.11.1
passlib==1.7.4
pyasn1==0.4.8
PyJWT==2.4.0
python-jose==3.3.0
